Rework popup design (unstable)
