+++ source/popup/actions/popupMenu.js
@@ -1,6 +1,6 @@
 import { createAction } from "redux-actions";
-import { POPUP_MENU_STATE_SHOW_ARCHIVES, POPUP_MENU_STATE_SHOW_OPTIONS, POPUP_MENU_STATE_TOGGLE } from "./types.js";
+// import { POPUP_MENU_STATE_SHOW_ARCHIVES, POPUP_MENU_STATE_SHOW_OPTIONS, POPUP_MENU_STATE_TOGGLE } from "./types.js";
 
-export const showArchivesMenu = createAction(POPUP_MENU_STATE_SHOW_ARCHIVES);
-export const showOptionsMenu = createAction(POPUP_MENU_STATE_SHOW_OPTIONS);
-export const toggleMenu = createAction(POPUP_MENU_STATE_TOGGLE);
+// export const showArchivesMenu = createAction(POPUP_MENU_STATE_SHOW_ARCHIVES);
+// export const showOptionsMenu = createAction(POPUP_MENU_STATE_SHOW_OPTIONS);
+// export const toggleMenu = createAction(POPUP_MENU_STATE_TOGGLE);
