+++ source/popup/actions/types.js
@@ -1,3 +1,3 @@
-export const POPUP_MENU_STATE_SHOW_ARCHIVES = "popupMenu/state/showArchives";
-export const POPUP_MENU_STATE_SHOW_OPTIONS = "popupMenu/state/showOptions";
-export const POPUP_MENU_STATE_TOGGLE = "popupMenu/toggleState";
+// export const POPUP_MENU_STATE_SHOW_ARCHIVES = "popupMenu/state/showArchives";
+// export const POPUP_MENU_STATE_SHOW_OPTIONS = "popupMenu/state/showOptions";
+// export const POPUP_MENU_STATE_TOGGLE = "popupMenu/toggleState";
