+++ source/popup/selectors/popupMenu.js
@@ -1,5 +1,5 @@
 const STATE_KEY = "popupMenu";
 
-export function getMenuState(state) {
-    return state[STATE_KEY].menuState;
-}
+// export function getMenuState(state) {
+//     return state[STATE_KEY].menuState;
+// }
